Add LightShader.java for reusable light code
